Remove "-debug" files
Those are just duplicates of the "release" versions. Probably
the "release" ones were supposed to be "minified". We can later on
do this as part of the msi build process, if we really need this.

At the same time, the FreeRDP_Logo.png image is in fact our background.
It's not used anywhere. A copy of it is included in "images/".